test: extract Common Handler test
